// packages/core-wagmi/src/types.ts
export type ConnectedChain = {
  id: Chain['id']
  namespace: Chain['namespace']
}
// ...
export type ChainStyle = {
  icon: string
  color: string
}

export type NotifyEventStyles = {
  backgroundColor: string
  borderColor: string
  eventIcon: string
  iconColor?: string
}